+++ src/main/java/ru/serega6531/packmate/model/pojo/StreamDto.java
@@ -15,7 +15,7 @@ public class StreamDto {
     private long endTimestamp;
     private Set<PatternDto> foundPatterns;
     private boolean favorite;
-    private byte ttl;
+    private int ttl;
     private String userAgentHash;
 
 }
